etl/requirements.txt
loguru==0.7.0
psycopg2-binary==2.9
pydantic==1.10.7
python-dotenv==1.0.0
elasticsearch==8.7.0